API/Reference

LuckyImaging.classic_lucky_image

LuckyImaging.classic_lucky_image!

LuckyImaging.fourier_lucky_image

LuckyImaging.fourier_lucky_image!

LuckyImaging.lucky_image

LuckyImaging.classic_lucky_image! — Method

classic_lucky_image!(out::AbstractMatrix, cube; dims, q, kwargs...)

Perform classic lucky imaging and store the combined frame in out. See classic_lucky_image for a full description.

LuckyImaging.classic_lucky_image — Method

classic_lucky_image(cube; dims, q, metric=:peak, register=:dft, kwargs...)

Perform classic lucky imaging for a cube of data along dimension dims. The selection fraction is set by the quantile q, where q=0.9 corresponds to a 10% selectrion fraction. The following options are available for modifying the metric used for selection as well as how the frames are registered.

Metrics

The following metrics, set with the keyword argument metric are available

:peak - the maximum value

:mean - the mean value

other - pass a function with the signature f(::AbstractMatrix) which returns a single value

Registration

The following options for registering the selected frames are available, set with the registerkeyword argument. Extra options can be provided to kwargs...

:dft - use SubpixelRegistration.jl to register the frames using their phase offsets. The reference frame will be the one with the highest metric, and keyword arguments like upsample_factor can be passed directly.

:peak - register to maximum value

:com - register to center of mass

Keyword arguments

dims - the dimension along which to perform lucky imaging (required)

q - the selection quantile (required)

metric - the metric used for selection. Default is :peak

register - the method used for registration. Default is :dft

window - if provided, will measure the metric and frame offsets inside a centered window with window side length.

kwargs... - additional keyword arguments will be passed to the register method (e.g., upsample_factor)

Examples

julia> cube = # load data ...

julia> res = classic_lucky_image(cube; dims=3, q=0.9, upsample_factor=10)

See Also

classic_lucky_image!

LuckyImaging.fourier_lucky_image! — Method

fourier_lucky_image!(out::AbstractMatrix, cube; dims, q, kwargs...)

Perform Fourier lucky imaging and store the combined frame in out. See fourier_lucky_image for a full descripton.

LuckyImaging.fourier_lucky_image — Method

fourier_lucky_image(cube; dims, q, register=:dft, kwargs...)

Perform Fourier lucky imaging following the algorithm in Garrel, Guyon, and Baudoz (2012).[1] This technique provides superior sharpness compared to classic lucky imaging for the same selection percentage. This means for a desired angular resolution (or Strehl ratio) a higher signal-to-noise ratio can be achieved.

Registration

Before processing the Fourier transform of the cube, the frames need coregistered to avoid phase ramps in the Fourier transform modulus. The following algorithms are available for registering the cube, set with the register keyword argument. Extra options can be provided to kwargs...

:dft - use SubpixelRegistration.jl to register the frames using their phase offsets. The reference frame will be the one with the highest metric, and keyword arguments like upsample_factor and refidx can be passed directly. By default the refidx will be set to the frame with the highest peak flux.

:peak - register to maximum value

:com - register to center of mass

Keyword arguments

dims - the dimension along which to perform lucky imaging (required)

q - the selection quantile (required)

register - the method used for registration. Default is :dft

maxfreq - if provided, will exclude frequencies higher than maxfreq * maximum(fftfreqs). In other words, a value of 1 uses the full Fourier transform, but a value of 0.5 will low-pass filter

kwargs... - additional keyword arguments will be passed to the register method (e.g., upsample_factor)

Examples

julia> cube = # load data ...

julia> res = fourier_lucky_image(cube; dims=3, q=0.5, upsample_factor=10);

See also

fourier_lucky_image!

LuckyImaging.lucky_image — Method

lucky_image(cube; dims, q, alg=:fourier, kwargs...)

Perform lucky imaging along dims with the chosen algorithm. If alg is :fourier fourier_lucky_image will be used, otherwise if it is :classic classic_lucky_image will be used. The keyword arguments for those methods can be passed directly through this method. q is the selection quantile; in other words, it is one minus the selection fraction.

Examples

julia> cube = # load data

julia> image = lucky_image(cube; dims=3, q=0.9, alg=:classic, upsample_factor=10);

julia> imagef = lucky_image(cube; dims=3, q=0.5, alg=:fourier, upsample_factor=10);

See also

classic_lucky_image, fourier_lucky_image

1Vincent Garrel et al 2012 "A Highly Efficient Lucky Imaging Algorithm: Image Synthesis Based on Fourier Amplitude Selection" PASP 124 861